+++ src/main/java/org/example/models/Transaction.java
@@ -4,18 +4,18 @@ import java.sql.Timestamp;
 
 public class Transaction {
     private final int id;
-    private final int amount;
+    private final double amount;
     private final Timestamp transactionDate;
     private final TransactionType transactionType;
 
-    public Transaction(int id, int amount, Timestamp transactionDate, TransactionType transactionType) {
+    public Transaction(int id, double amount, Timestamp transactionDate, TransactionType transactionType) {
         this.id = id;
         this.amount = amount;
         this.transactionDate = transactionDate;
         this.transactionType = transactionType;
     }
 
-    public Transaction(int amount, Timestamp transactionDate, TransactionType transactionType) {
+    public Transaction(double amount, Timestamp transactionDate, TransactionType transactionType) {
         this.id = 0;
         this.amount = amount;
         this.transactionDate = transactionDate;
@@ -26,7 +26,7 @@ public class Transaction {
         return id;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
